Refactoring the HTML cleanup plugin
