+++ .gitignore
@@ -8,3 +8,4 @@ yarn-error.log
 .DS_Store
 .vs
 /types/
+/JODIT-DEVELOPMENT-GUIDE.md
