+++ .npmignore
@@ -19,3 +19,5 @@ test.html
 .stylelintrc
 app.css
 composer.json
+/JODIT-DEVELOPMENT-GUIDE.md
+/test.html
