+++ src/core/helpers/html/index.ts
@@ -12,4 +12,5 @@ export * from './apply-styles';
 export * from './clean-from-word';
 export * from './htmlspecialchars';
 export * from './strip-tags';
+export * from './safe-html';
 export * from './nl2br';
