+++ src/plugins/fix/clean-html/helpers/remove-inv-text-nodes.ts
@@ -1,0 +1,44 @@
+/*!
+ * Jodit Editor (https://xdsoft.net/jodit/)
+ * Released under MIT see LICENSE.txt in the project root for license information.
+ * Copyright (c) 2013-2022 Valeriy Chupurnov. All rights reserved. https://xdsoft.net
+ */
+
+import type { IJodit } from 'jodit/types';
+import { INVISIBLE_SPACE_REG_EXP as INV_REG } from 'jodit/core/constants';
+import { Dom } from 'jodit/core/dom/dom';
+
+export function removeInvTextNodes(jodit: IJodit): void {
+	const currentNode = jodit.s.current();
+
+	if (!currentNode) {
+		return;
+	}
+
+	const currentParagraph = Dom.up(currentNode, Dom.isBlock, jodit.editor);
+
+	if (!currentParagraph) {
+		return;
+	}
+
+	Dom.each(currentParagraph, node => {
+		if (!Dom.isText(node) || node.nodeValue == null) {
+			return;
+		}
+
+		if (
+			INV_REG().test(node.nodeValue) &&
+			node.nodeValue.replace(INV_REG(), '').length !== 0
+		) {
+			node.nodeValue = node.nodeValue.replace(INV_REG(), '');
+
+			if (node === currentNode && jodit.s.isCollapsed()) {
+				jodit.s.setCursorAfter(node);
+			}
+
+			if (!node.nodeValue) {
+				Dom.safeRemove(node);
+			}
+		}
+	});
+}
