+++ src/plugins/fix/clean-html/helpers/visitor/cases/allow-attributes.ts
@@ -1,0 +1,44 @@
+/*!
+ * Jodit Editor (https://xdsoft.net/jodit/)
+ * Released under MIT see LICENSE.txt in the project root for license information.
+ * Copyright (c) 2013-2022 Valeriy Chupurnov. All rights reserved. https://xdsoft.net
+ */
+
+import type { IDictionary, IJodit } from 'jodit/types';
+import { Dom } from 'jodit/core/dom/dom';
+
+/**
+ * @private
+ */
+export function allowAttributes(
+	jodit: IJodit,
+	nodeElm: Node,
+	hasChanges: boolean,
+	allow: IDictionary | false
+): boolean {
+	if (allow && Dom.isElement(nodeElm) && allow[nodeElm.nodeName] !== true) {
+		const attrs: NamedNodeMap = (nodeElm as Element).attributes;
+
+		if (attrs && attrs.length) {
+			const removeAttrs: string[] = [];
+
+			for (let i = 0; i < attrs.length; i += 1) {
+				const attr = allow[nodeElm.nodeName][attrs[i].name];
+
+				if (!attr || (attr !== true && attr !== attrs[i].value)) {
+					removeAttrs.push(attrs[i].name);
+				}
+			}
+
+			if (removeAttrs.length) {
+				hasChanges = true;
+			}
+
+			removeAttrs.forEach(attr => {
+				(nodeElm as Element).removeAttribute(attr);
+			});
+		}
+	}
+
+	return hasChanges;
+}
